refactor: web 문서화 구조 리펙토링
