+++ hashtagmap-web/src/test/java/com/songpapeople/hashtagmap/api/MapApiControllerTest.java
@@ -8,10 +8,13 @@ import com.songpapeople.hashtagmap.place.domain.model.Category;
 import com.songpapeople.hashtagmap.place.domain.model.Location;
 import com.songpapeople.hashtagmap.place.domain.model.Place;
 import com.songpapeople.hashtagmap.place.domain.model.Point;
+import com.songpapeople.hashtagmap.service.InstagramQueryService;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 
 import java.util.Arrays;
@@ -22,8 +25,12 @@ import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuild
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+@WebMvcTest(controllers = MapApiController.class)
 public class MapApiControllerTest extends MapApiDocumentation {
     private List<MarkerResponse> markerResponses;
+
+    @MockBean
+    protected InstagramQueryService instagramQueryService;
 
     @BeforeEach
     void setUp() {
