+++ hashtagmap-web/src/test/java/com/songpapeople/hashtagmap/docs/ApiDocument.java
@@ -1,13 +1,9 @@
 package com.songpapeople.hashtagmap.docs;
 
-import com.songpapeople.hashtagmap.service.InstagramPostQueryService;
-import com.songpapeople.hashtagmap.service.InstagramQueryService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.restdocs.RestDocumentationContextProvider;
 import org.springframework.restdocs.RestDocumentationExtension;
 import org.springframework.test.web.servlet.MockMvc;
@@ -17,17 +13,10 @@ import org.springframework.web.context.WebApplicationContext;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.documentationConfiguration;
 
 @ExtendWith(RestDocumentationExtension.class)
-@WebMvcTest
 @AutoConfigureMockMvc
 public class ApiDocument {
     @Autowired
     protected MockMvc mockMvc;
-
-    @MockBean
-    protected InstagramPostQueryService instagramPostQueryService;
-
-    @MockBean
-    protected InstagramQueryService instagramQueryService;
 
     @BeforeEach
     public void setUp(WebApplicationContext webApplicationContext,
